Apply Metered License, Retrieve Parent Shape and Verify if a shape is in group using Aspose.3D for .NET 17.02.0
The ParentShape property and IsInGroup method is added in the Shape
class. ParentShape property in the Shape class allows to retrieve the
parent shape. IsInGroup method in the Shape class allows to verify
whether a shape is part of a group. The new Metered class allows to
apply metered license.